+++ src/firstGradeFunctionSearcher.py
@@ -1,43 +1,4 @@
-import geneticPie
-from geneticPie import *
-import random
-from random import *
-
-class FloatGen(ValuableGen):
-
-    #public
-    def __init__(self, *param):
-        ValuableGen.__init__(self, param[0])
-        self.default = self.value
-
-    #public
-    def get_req_gens(self):
-        return {}
-
-    #public
-    def is_mutable(self, *param):
-        return True
-
-    #public
-    def mutate(self, *param):
-        percent = int(random() * 100)
-        mutatio = random() + 0.5
-        if percent <= 9:
-            self.value = {
-                0 : self.default,
-                1 : self.value - mutatio,
-                2 : self.value + mutatio,
-                3 : self.value * mutatio,
-                4 : self.value / mutatio,
-                5 : int(self.value),
-                6 : int(self.value) + 0.5,
-                7 : int(self.value) - 0.5,
-                8 : int(self.value) + 0.1,
-                9 : int(self.value) - 0.1
-            }[percent]
-
-    def new_instace(self):
-        return self.__class__(self.value)
+from defaultGens import *
 
 class FuncFGen(RunnableGen):
 
@@ -81,7 +42,7 @@ class FuncFInd(Individual):
 
         return total
 
-pars = [(0, 0), (20,20)]
+pars = [(3, 1), (6, 2)]
 sim = Simulation()
 sim.population.append(FuncFInd(None))
 
@@ -97,8 +58,5 @@ for k in range(5000):
     for k in range(5):
         sim.population[k+5] = sim.population[k].new_instace()
 
-        
-
-#print map((lambda x : (x.calculate_fitness((4, 4)))),sim.population)
 for ind in sim.population:
     print ind.calculate_fitness(*pars), 'a = ', ind.gens['a'].value, 'b = ', ind.gens['b'].value
